feat: impl secret watching and refactor controller tests - Watch Secret resources to trigger HeliosApp reconciliation on token updates - Merge unit tests into heliosapp_controller_test.go for unified mocking - Update testing docs
